+++ R/compute_tax_distinct_formula.R
@@ -4,7 +4,6 @@ compute_tax_distinct_formula <- function(x, y) {
 
   temp <- names(y) %in% x$scientificName
   tax_hier_temp <- y[c(temp)]
-  print(length(tax_hier_temp))
   n_spec <- length(tax_hier_temp)
 
   if (n_spec < 3) {